nsis: 3.06.1 -> 3.11

+++ pkgs/misc/jackaudio/default.nix
@@ -1,22 +1,36 @@
-{ lib, stdenv, fetchFromGitHub, fetchpatch2, pkg-config, python3Packages, makeWrapper
-, libsamplerate, libsndfile, readline, eigen, celt
-, wafHook
-# Darwin Dependencies
-, aften, AudioUnit, CoreAudio, libobjc, Accelerate
+{
+  lib,
+  stdenv,
+  fetchFromGitHub,
+  fetchpatch2,
+  pkg-config,
+  python3Packages,
+  makeWrapper,
+  libsamplerate,
+  celt,
+  wafHook,
+  # Darwin Dependencies
+  aften,
 
-# Optional Dependencies
-, dbus ? null, libffado ? null, alsa-lib ? null
-, libopus ? null
+  # BSD Dependencies
+  freebsd,
 
-# Extra options
-, prefix ? ""
+  # Optional Dependencies
+  dbus ? null,
+  libffado ? null,
+  alsa-lib ? null,
+  libopus ? null,
 
-, testers
+  # Extra options
+  prefix ? "",
+
+  testers,
 }:
 
 let
   inherit (python3Packages) python dbus-python;
-  shouldUsePkg = pkg: if pkg != null && lib.meta.availableOn stdenv.hostPlatform pkg then pkg else null;
+  shouldUsePkg =
+    pkg: if pkg != null && lib.meta.availableOn stdenv.hostPlatform pkg then pkg else null;
 
   libOnly = prefix == "lib";
 
@@ -37,14 +51,32 @@ stdenv.mkDerivation (finalAttrs: {
     sha256 = "sha256-Cslfys5fcZDy0oee9/nM5Bd1+Cg4s/ayXjJJOSQCL4E=";
   };
 
-  outputs = [ "out" "dev" ];
-
-  nativeBuildInputs = [ pkg-config python makeWrapper wafHook ];
-  buildInputs = [ libsamplerate libsndfile readline eigen celt
-    optDbus optPythonDBus optLibffado optAlsaLib optLibopus
-  ] ++ lib.optionals stdenv.hostPlatform.isDarwin [
-    aften AudioUnit CoreAudio Accelerate libobjc
+  outputs = [
+    "out"
+    "dev"
   ];
+
+  nativeBuildInputs = [
+    pkg-config
+    python
+    wafHook
+  ] ++ lib.optionals (optDbus != null) [ makeWrapper ];
+  buildInputs =
+    [
+      libsamplerate
+      celt
+      optDbus
+      optPythonDBus
+      optLibffado
+      optAlsaLib
+      optLibopus
+    ]
+    ++ lib.optionals stdenv.hostPlatform.isDarwin [
+      aften
+    ]
+    ++ lib.optionals stdenv.hostPlatform.isFreeBSD [
+      freebsd.libsysinfo
+    ];
 
   patches = [
     (fetchpatch2 {
@@ -59,33 +91,43 @@ stdenv.mkDerivation (finalAttrs: {
     patchShebangs --build svnversion_regenerate.sh
   '';
 
-  wafConfigureFlags = [
-    "--classic"
-    "--autostart=${if (optDbus != null) then "dbus" else "classic"}"
-  ] ++ lib.optional (optDbus != null) "--dbus"
+  wafConfigureFlags =
+    [
+      "--classic"
+      "--autostart=${if (optDbus != null) then "dbus" else "classic"}"
+    ]
+    ++ lib.optional (optDbus != null) "--dbus"
     ++ lib.optional (optLibffado != null) "--firewire"
-    ++ lib.optional (optAlsaLib != null) "--alsa";
+    ++ lib.optional (optAlsaLib != null) "--alsa"
+    ++ lib.optional (
+      stdenv.hostPlatform != stdenv.buildPlatform
+    ) "--platform=${stdenv.hostPlatform.parsed.kernel.name}";
 
-  postInstall = (if libOnly then ''
-    rm -rf $out/{bin,share}
-    rm -rf $out/lib/{jack,libjacknet*,libjackserver*}
-  '' else lib.optionalString (optDbus != null) ''
-    wrapProgram $out/bin/jack_control --set PYTHONPATH $PYTHONPATH
-  '');
+  postInstall = (
+    if libOnly then
+      ''
+        rm -rf $out/{bin,share}
+        rm -rf $out/lib/{jack,libjacknet*,libjackserver*}
+      ''
+    else
+      lib.optionalString (optDbus != null) ''
+        wrapProgram $out/bin/jack_control --set PYTHONPATH $PYTHONPATH
+      ''
+  );
 
   postFixup = ''
     substituteInPlace "$dev/lib/pkgconfig/jack.pc" \
-      --replace "$out/include" "$dev/include"
+      --replace-fail "$out/include" "$dev/include"
   '';
 
   passthru.tests.pkg-config = testers.testMetaPkgConfig finalAttrs.finalPackage;
 
-  meta = with lib; {
+  meta = {
     description = "JACK audio connection kit, version 2 with jackdbus";
     homepage = "https://jackaudio.org";
-    license = licenses.gpl2Plus;
+    license = lib.licenses.gpl2Plus;
     pkgConfigModules = [ "jack" ];
-    platforms = platforms.unix;
+    platforms = lib.platforms.unix;
     maintainers = [ ];
   };
 })
