+++ pkgs/tools/text/gawk/default.nix
@@ -1,30 +1,35 @@
-{ lib, stdenv, fetchurl
-, removeReferencesTo
-, runtimeShellPackage
-# TODO: links -lsigsegv but loses the reference for some reason
-, withSigsegv ? (false && stdenv.hostPlatform.system != "x86_64-cygwin"), libsigsegv
-, interactive ? false, readline
-, autoreconfHook # no-pma fix
+{
+  lib,
+  stdenv,
+  fetchurl,
+  removeReferencesTo,
+  runtimeShellPackage,
+  texinfo,
+  interactive ? false,
+  readline,
+  autoreconfHook, # no-pma fix
 
-/* Test suite broke on:
-       stdenv.hostPlatform.isCygwin # XXX: `test-dup2' segfaults on Cygwin 6.1
-    || stdenv.hostPlatform.isDarwin # XXX: `locale' segfaults
-    || stdenv.hostPlatform.isSunOS  # XXX: `_backsmalls1' fails, locale stuff?
-    || stdenv.hostPlatform.isFreeBSD
-*/
-, doCheck ? (interactive && stdenv.hostPlatform.isLinux), glibcLocales ? null
-, locale ? null
+  /*
+    Test suite broke on:
+        stdenv.hostPlatform.isCygwin # XXX: `test-dup2' segfaults on Cygwin 6.1
+     || stdenv.hostPlatform.isDarwin # XXX: `locale' segfaults
+     || stdenv.hostPlatform.isSunOS  # XXX: `_backsmalls1' fails, locale stuff?
+     || stdenv.hostPlatform.isFreeBSD
+  */
+  doCheck ? (interactive && stdenv.hostPlatform.isLinux),
+  glibcLocales ? null,
+  locale ? null,
 }:
 
 assert (doCheck && stdenv.hostPlatform.isLinux) -> glibcLocales != null;
 
 stdenv.mkDerivation rec {
   pname = "gawk" + lib.optionalString interactive "-interactive";
-  version = "5.2.2";
+  version = "5.3.2";
 
   src = fetchurl {
     url = "mirror://gnu/gawk/gawk-${version}.tar.xz";
-    hash = "sha256-PB/OFEa0y+4c0nO9fsZLyH2J9hU3RxzT4F4zqWWiUOk=";
+    hash = "sha256-+MNIZQnecFGSE4sA7ywAu73Q6Eww1cB9I/xzqdxMycw=";
   };
 
   # PIE is incompatible with the "persistent malloc" ("pma") feature.
@@ -34,33 +39,43 @@ stdenv.mkDerivation rec {
   hardeningDisable = [ "pie" ];
 
   # When we do build separate interactive version, it makes sense to always include man.
-  outputs = [ "out" "info" ]
-    ++ lib.optional (!interactive) "man";
+  outputs = [
+    "out"
+    "info"
+  ] ++ lib.optional (!interactive) "man";
 
   strictDeps = true;
 
   # no-pma fix
-  nativeBuildInputs = [
-    autoreconfHook
-  ] ++ lib.optionals interactive [
-    removeReferencesTo
-  ] ++ lib.optionals (doCheck && stdenv.hostPlatform.isLinux) [
-    glibcLocales
-  ];
+  nativeBuildInputs =
+    [
+      autoreconfHook
+      texinfo
+    ]
+    ++ lib.optionals interactive [
+      removeReferencesTo
+    ]
+    ++ lib.optionals (doCheck && stdenv.hostPlatform.isLinux) [
+      glibcLocales
+    ];
 
-  buildInputs = lib.optionals interactive [
-    runtimeShellPackage
-    readline
-  ] ++ lib.optionals withSigsegv [
-    libsigsegv
-  ] ++ lib.optionals stdenv.hostPlatform.isDarwin [
-    locale
-  ];
+  buildInputs =
+    lib.optionals interactive [
+      runtimeShellPackage
+      readline
+    ]
+    ++ lib.optionals stdenv.hostPlatform.isDarwin [
+      locale
+    ];
 
   configureFlags = [
-    (if withSigsegv then "--with-libsigsegv-prefix=${libsigsegv}" else "--without-libsigsegv")
     (if interactive then "--with-readline=${readline.dev}" else "--without-readline")
   ];
+
+  env = lib.optionalAttrs stdenv.hostPlatform.isDarwin {
+    # TODO: figure out a better way to unbreak _NSGetExecutablePath invocations
+    NIX_CFLAGS_COMPILE = "-Wno-implicit-function-declaration";
+  };
 
   makeFlags = [
     "AR=${stdenv.cc.targetPrefix}ar"
@@ -68,21 +83,24 @@ stdenv.mkDerivation rec {
 
   inherit doCheck;
 
-  postInstall = (if interactive then ''
-    remove-references-to -t "$NIX_CC" "$out"/bin/gawkbug
-    patchShebangs --host "$out"/bin/gawkbug
-  '' else ''
-    rm "$out"/bin/gawkbug
-  '') + ''
-    rm "$out"/bin/gawk-*
-    ln -s gawk.1 "''${!outputMan}"/share/man/man1/awk.1
-  '';
+  postInstall =
+    (
+      if interactive then
+        ''
+          remove-references-to -t "$NIX_CC" "$out"/bin/gawkbug
+          patchShebangs --host "$out"/bin/gawkbug
+        ''
+      else
+        ''
+          rm "$out"/bin/gawkbug
+        ''
+    )
+    + ''
+      rm "$out"/bin/gawk-*
+      ln -s gawk.1 "''${!outputMan}"/share/man/man1/awk.1
+    '';
 
-  passthru = {
-    libsigsegv = if withSigsegv then libsigsegv else null; # for stdenv bootstrap
-  };
-
-  meta = with lib; {
+  meta = {
     homepage = "https://www.gnu.org/software/gawk/";
     description = "GNU implementation of the Awk programming language";
     longDescription = ''
@@ -98,9 +116,9 @@ stdenv.mkDerivation rec {
       makes it possible to handle many data-reformatting jobs with just a few
       lines of code.
     '';
-    license = licenses.gpl3Plus;
-    platforms = platforms.unix ++ platforms.windows;
-    maintainers = [ ];
+    license = lib.licenses.gpl3Plus;
+    platforms = lib.platforms.unix ++ lib.platforms.windows;
+    teams = [ lib.teams.helsinki-systems ];
     mainProgram = "gawk";
   };
 }
